Add model data account view
